server/requirements.txt
anyio==3.7.1
attrs==23.1.0
bcrypt==4.0.1
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
click-log==0.4.0
colorama==0.4.6
confection==0.1.0
cryptography==41.0.2
cymem==2.0.7
deprecation==2.1.0
docutils==0.20.1
docx2txt==0.8
dotty-dict==1.3.1
easyocr==1.7.0
filelock==3.12.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
fsspec==2023.6.0
future==0.18.3
geonamescache==2.0.0
geotext==0.4.0
gitdb==4.0.10
GitPython==3.1.32
gotrue==1.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==24.2.0
langcodes==3.3.0
lazy_loader==0.3
MarkupSafe==2.1.3
mime==0.1.0
more-itertools==10.0.0
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numpy==1.25.1
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
pathy==0.10.2
pdfminer==20191125
pdfminer.six==20221105
Pillow==10.0.0
pkginfo==1.9.6
plac==1.1.3
postgrest==0.10.6
preshed==3.0.8
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
PyPDF2==3.0.1
pyrsistent==0.19.3
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
readme-renderer==40.0
realtime==1.0.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.9.2
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
semver==2.13.0
shapely==2.0.1
six==1.16.0
sklearn==0.0.post7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.6.0
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.5
srsly==2.4.7
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
sympy==1.12
thinc==8.1.10
threadpoolctl==3.2.0
tifffile==2023.7.18
tokenizers==0.13.3
tomlkit==0.12.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
twine==3.8.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==0.10.1
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.6
zipp==3.16.2
